requirements.txt
Flask==2.3.3
requests==2.31.0
werkzeug
langchain-community
langchain-huggingface
faiss-cpu
PyPDF2
docx2txt
langchain
langchain-openai
pypdf
psycopg2-binary>=2.9
python3-saml
pyJWT
